Add FFTW-QFT functions to public interface
The new functions for forwards and backwards transformations explicitly
mention `FFTW` and take the first and last qubit of a contiguous range
which form a sub-register. These functions will likely disappear in the
future when internal routines decide automatically whether to call FFTW
directly, re-order qubits first, or fall back on the gate-based
implementation.